Revert "Merge pull request #2 from zhaarey/codex/change-itunesalbumid-and-itunesartistid-to-int64"
This reverts commit ba4505156fad3b0bb19e70f65947b94b38b977fb, reversing
changes made to 2c70f6b1bf76e8b847278abee52198fcd170dabd.

+++ objects.go
@@ -294,8 +294,8 @@ type MP4Tags struct {
 	DiscTotal int16
 	Genre Genre
 	ItunesAdvisory ItunesAdvisory
-	ItunesAlbumID int64
-	ItunesArtistID int64
+	ItunesAlbumID int32
+	ItunesArtistID int32
 	Lyrics string
 	Narrator string
 	OtherCustom map[string][]string
